Extract shared predicates into PAGXAnalyzer and deduplicate test utilities.
